+++ tomcat/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,8 +1,9 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
 import java.util.Map;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
+import org.apache.catalina.controller.AbstractController;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.request.Request;
 import org.apache.coyote.http11.message.response.Response;
