+++ tomcat/src/main/java/nextstep/jwp/controller/RootController.java
@@ -1,6 +1,7 @@
-package org.apache.coyote.http11.controller;
+package nextstep.jwp.controller;
 
-import org.apache.coyote.http11.exception.UnsupportedRequestException;
+import org.apache.catalina.controller.AbstractController;
+import org.apache.catalina.exception.UnsupportedRequestException;
 import org.apache.coyote.http11.message.ContentType;
 import org.apache.coyote.http11.message.HttpStatus;
 import org.apache.coyote.http11.message.request.Request;
