+++ tomcat/src/main/java/org/apache/catalina/controller/ControllerMapper.java
@@ -1,0 +1,23 @@
+package org.apache.catalina.controller;
+
+import java.util.HashMap;
+import java.util.Map;
+import nextstep.jwp.controller.StaticFileController;
+import org.apache.coyote.http11.message.request.RequestLine;
+
+public class ControllerMapper {
+
+    private static final Controller DEFAULT_CONTROLLER = new StaticFileController();
+    private final Map<ControllerStatus, Controller> handlers = new HashMap<>();
+
+    public ControllerMapper() {
+    }
+
+    public void addController(ControllerStatus controllerStatus, Controller controller) {
+        handlers.put(controllerStatus, controller);
+    }
+
+    public Controller findHandler(RequestLine requestLine) {
+        return handlers.getOrDefault(ControllerStatus.from(requestLine), DEFAULT_CONTROLLER);
+    }
+}
