+++ tomcat/src/main/java/org/apache/catalina/exception/file/NoExistFileExtensionException.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.exception.file;
+package org.apache.catalina.exception.file;
 
 public class NoExistFileExtensionException extends RuntimeException {
